+++ controllers/eks_dns_test.go
@@ -494,7 +494,7 @@ var _ = Describe("Dns Zone reconciler", func() {
 
 						BeforeEach(func() {
 							awsManagedControlPlane.Annotations = map[string]string{
-								controllers.AWSDNSHostedZoneName: customHostedZoneName,
+								controllers.DNSHostedZoneNameAnnotation: customHostedZoneName,
 							}
 							route53Client.GetHostedZoneIdByNameReturns("custom-parent-hosted-zone-id", nil)
 							route53Client.CreateHostedZoneReturns("hosted-zone-id", nil)
@@ -528,7 +528,7 @@ var _ = Describe("Dns Zone reconciler", func() {
 
 						BeforeEach(func() {
 							awsManagedControlPlane.Annotations = map[string]string{
-								controllers.AWSDNSDelegationRoleARN: customDelegationRoleARN,
+								controllers.AWSDNSDelegationRoleARNAnnotation: customDelegationRoleARN,
 							}
 							route53Client.GetHostedZoneIdByNameReturns("parent-hosted-zone-id", nil)
 							route53Client.CreateHostedZoneReturns("hosted-zone-id", nil)
@@ -543,7 +543,7 @@ var _ = Describe("Dns Zone reconciler", func() {
 					When("a custom hosted zone name without valid parent is specified", func() {
 						BeforeEach(func() {
 							awsManagedControlPlane.Annotations = map[string]string{
-								controllers.AWSDNSHostedZoneName: "invalid",
+								controllers.DNSHostedZoneNameAnnotation: "invalid",
 							}
 							route53Client.CreateHostedZoneReturns("hosted-zone-id", nil)
 						})